+++ src/formfactors/index.ts
@@ -1,4 +1,5 @@
 import { ChatbotFormfactorDefinition } from './chatbot/ChatbotFormfactorDefinition';
+import { CompletionFormfactorDefinition } from './completion/CompletionFormfactorDefinition';
 import { GeneratorFormfactorDefinition } from './generator/GeneratorFormfactorDefinition';
 import { GenericFormfactorDefinition } from './generic/GenericFormfactorDefinition';
 import { ImageGeneratorFormfactorDefinition } from './image-generator/ImageGeneratorFormfactorDefinition';
@@ -19,6 +20,7 @@ export const FORMFACTOR_DEFINITIONS = [
     MatcherFormfactorDefinition,
     GeneratorFormfactorDefinition,
     ImageGeneratorFormfactorDefinition,
+    CompletionFormfactorDefinition,
     // <- [🛬] When making new formfactor, copy the _boilerplate and link it here
 ] as const;